move direction selector up to the bridge
