+++ src/components/Bridge/index.ts
@@ -1,0 +1,2 @@
+import Bridge from './Bridge'
+export default Bridge
